Resolves: tdf#119739 switch date field input to NF_EVALDATEFORMAT_FORMAT_INTL
To evaluate date input against date acceptance patterns in the
order first format locale, if any, then formatter (system) locale.

The slight drawback is when the field has no content there is no
indication of that the format's locale's patterns would be
preferred, whereas the old NF_EVALDATEFORMAT_INTL_FORMAT preferred
the default locale's patterns.

The big advantage is that editing a date actually uses the
patterns associated with the presented display string.

Change-Id: Id3f11e26a0e32b95f26b69a681433deb87b7c6b0
Reviewed-on: https://gerrit.libreoffice.org/61726
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins