sw_redlinehide_3: remove number parameter from SwDoc::SetCurFootnote
The number is pretty useless: none of the callers know what the proper
auto-counted number is going to be, so they either pass in the number
that the footnote already has, or 0.

It is particularly funny if the SwPaM is actually a range containing
multiple footnotes; fortunately such a call is apparently not possible
currently.

Just remove the parameter and use the existing footnote number directly.

Change-Id: I45dee999b8b0b630eff4347a563ccca2869f84cc