Special methods NewStandardLayer and SetStandardLayer not needed
The method SetStandardLayer() is only used once, in method
SdrLayerAdmin::NewStandardLayer(). This method is only used once, in
OReportDefinition::init(). The difference between a layer 'Standard'
and other layers is only in mbUserDefinedLayer. But this difference is
no where evaluated. Therefore an ordinary layer is sufficient. The
layer name is neither stored nor used in the UI, thus translation is
not needed. A layer name must be set, but in this case it is not
used in code, thus handle it same as for layer 'back'.

Change-Id: Ifd4400d432d3b4702e950d3a717ed13d4075c279
Reviewed-on: https://gerrit.libreoffice.org/62144
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>