Look for the generated native-code.h where it now is
The makefile here was changed already some weeks ago to put
native-code.h in workdir and not in the source directory, but I forgot
to make sure this file still compiled, as it is used by
LibreOfficeLight only.

Yes, it is ugly to use the workdir/CustomTarget/ios pathname, so sue
me.

Change-Id: I568d933c1d1384041632f432053d0a0c64c485c2