tdf#118244 pdfexport: radio buttons use groupname now
The previous implementation grouped radio buttons if their
object name was the same. Likely this is a very old implementation,
because the current radio buttons have a groupname property which
links them together (although that too needed fixing in doc/docx),
and their object names are unique.

The old implementation still works - so that still needs to be
supported, but I think I'll do that in a separate patch, so
that it can be easily reverted if the old implementation is
deprecated.

Edge cases tested:
-groupID of 0 works fine - doesn't have to be 1-based.
-empty group name works fine (but breaks the old impl).
-writer, calc,

Change-Id: I84aebdac18b9edfa5ffcbfb23c15d0f37fcd47d1
Reviewed-on: https://gerrit.libreoffice.org/62742
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>