We now use a fixed name ICU data file in the iOS app
That makes it much easier to use the same Xcode project file when
building against different versions of LibreOffice, with different ICU
data files. The configure script in the "online" directory creates a
ICU.dat symlink to the ICU data file (with a name like icudt60l.dat in
a 6.0 branch, or icudt63l.dat in a current master) in the LibreOffice
build directory, and that is what gets copied into the app bundle.

Change-Id: Ibe2ca85f66e2d4973d6ba3c52fc4d60e113d8f9a