For iOS, do actually export UI builder factory functions
In the DISABLE_DYNLOADING case there is normally no need for functions
marked with SAL_DLLPUBLIC_EXPORT to be exported, as these functions
won't be dynamically looked up anyway. Thus, when DISABLE_DYNLOADING,
SAL_DLLPUBLIC_EXPORT is defined in <sal/types.h> to actually mean
__attribute__ ((visibility("hidden"))).

But we do need to export the UI builder factory functions so that the
osl_getFunctionSymbol() in VclBuilder::makeObject() finds them.

(I kinda dislike looking up symbols with dlsym() from the same binary.
We know that the function is there and what its name is, we could just
call it directly. But makeObject() gets the function name as a string,
so we would need a long set of string comparisons to select which
function to call. A bit ugly. Let's see if I can come up with
something elegant enough later.)

Change-Id: Idceaf8c1ed54cd7d372bf4eb85d0428f9b57baeb
Reviewed-on: https://gerrit.libreoffice.org/62799
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>