tdf120965 Simplify SdrPathObj::AddToHdlList()
A polygon consists of line segments without controls and curve
segments with two controls. Where two segments meet, can only be
'line to line', 'line to curve' or 'curve to curve'. So for a
handle which marks a meeting point of two segments, there are 0
or 1 control before and 0 or 1 control after. Thus a loop is not
needed and makes the code harder to read. Therefore I have
removed the loop.

Change-Id: I48bcdbdf028542e08b3f5ed3f66907f2993e26bd
Reviewed-on: https://gerrit.libreoffice.org/63036
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>