Reorder ScFormulaCell members
When evaluating/getting result from a formulacell(which
has been evaluated before), first bDirty/bTableOpDirty
are accessed from MaybeInterpret()/NeedsInterpret(), then
aResult is accessed, so it makes sense to put the first
7 bools(with bDirty/bTableOpDirty) in the beginning, then
aResult, then rest of bools and other remaining members. The
expectation is that when bDirty is accessed in NeedsInterpret(),
the relevant contents of aResult is also fetched to cache
and remains there till aResult is needed.

Change-Id: Ide0d1e809ca045ef347520f26c4d3ab9faa994bb
Reviewed-on: https://gerrit.libreoffice.org/63065
Tested-by: Jenkins
Reviewed-by: Dennis Francis <dennis.francis@collabora.com>