LOK: disable LOKit in unit-tests only after destroying all
When we disable LOKit at the end of a unit-test, we do so
before any of the smart-pointers go out of scope. These
instances are then destroyed when LOKit is no longer active,
in spite of the fact that they were created while it was.

This leaves some dialog/window routing notifiers dangling,
which cause assertion failures and access-after-free (when
they are invoked on destruction inadvertently). The latter
issue was found with 'make CppunitTest_sd_tiledrendering
VALGRIND=memcheck'.

The fix is to move these flags to the setup and teardown
stages of cpp-unit.

Change-Id: Ic4481305ca1a5cca14d70a4ebbc820a3fdeea9cc
Reviewed-on: https://gerrit.libreoffice.org/80028
Tested-by: Jenkins
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>