sw_redlinehide_3: add second result to SwGetExpField, SwSetExpField
This is quite tricky & somewhat ugly due to the SwValueField base class
& we need to duplicate its member too :(

It would be possible to filter the deleted fields in
SwDocUpdateField::MakeFieldList(), but the problem with that is that
the instance is long-lived & might be used with different layout
settings, so it's probably better to filter in the 5 or so client
functions that iterate over the GetSortList().

Note that this will also filter other fields including database fields,
hopefully that shouldn't cause problems.

Change-Id: I0cdbc7757fa529598b5dbceec0bd2c4a619be05a