Make Application::SetDialogCancelMode "private"
Change callers to Application::EnableHeadlessMode instead except
for the Frame dispose sequence, which has to disable a fatal
mode temporary.

Since Frame is not a unique identifier, rename it to XFrameImpl,
otherwise we get a "non-unique" compile error for other files.

When we update our gcc baseline, we should be able to forward
declare and friend XFrameImpl and really make it private.

Change-Id: Ifcad0355adade8f6539ec092b174dd70af8352c2
Reviewed-on: https://gerrit.libreoffice.org/64050
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>