Remove obsolete CLANG_VERSION checks
...now that the (Linux) Clang baseline has been bumped ot 5.0.2 (see
<https://lists.freedesktop.org/archives/libreoffice/2018-November/081435.html>
"minutes of ESC call ...").

(This commit precedes a commit that will actually enforce Clang >= 5.0.2 in
configure.ac.  However, it was easier to do it in this order, as that other
commit will drop CLANG_VERSION when building with Apple Clang (where version
numbers are different from upstream), so these checks would start to fail
there.)

Change-Id: Icb3df0d0cf476e14c5453f02bdfc9e5bed066ca1
Reviewed-on: https://gerrit.libreoffice.org/63898
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>