sw_redlinehide_4a: fix asserts caused by invalid index in SwUpdateAttr
There is no reason for SwpHints::TryInsertHint() to add 1 to the end
index of an empty hint, which results in an invalid index when the end
of the hint is at the end of the node.

The only user of the indexes in SwUpdateAttr is
SwTextFrame::SwClientNotify() and it already extends a 0 range to 1.

Change-Id: I7355bac931cd8dd390f96d04546d73273f364cda