oox smartart, accent process: adjust size of connector from constraints
The constraints explicitly said that the width should be larger than the
height, but it was the opposite as constraints were not parsed.

Unfortunately it would be too brave for globally start handling all
constraints which lack a forName, so add a switch to opt in for this,
and use that with the conn algorithm. All clients should migrate to
bRequireForName=true at some stage, though.

Change-Id: I24ae79b141c0f7a11e4d19f141759fc1dd2169b0
Reviewed-on: https://gerrit.libreoffice.org/64350
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins