tdf#121815 just use valid cached layout glyphs
I'm not exacly sure when and why the font cache is invalidated,
but as a result this also invalidates the cached layouted glyphs.

So just check the glyph layout cache state before trying to use
it. But actually the status bar should update it's cache, so I'm
not sure where to really put the fix. At least this won't crash
LO anymore.

Change-Id: I5f3ff8b9d56808af74d1419e878819d6019cd893
Reviewed-on: https://gerrit.libreoffice.org/64529
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>