tdf#69569 implement proposed change to ODF1.2 part 2 §4.11.7.7
Calculating the number of days in a year in AddIn function GetDaysInYear()
for basis 1 (actual/actual) gave wrong results.
Now the results are correct and the same as in Excel.

Extended the unit test document for function YEARFRAC and corrected the
unit test document for function AMORDEGRC (verified the results with Excel).

Change-Id: Ic68f108496f41dec71b3616095dff80512a64c31
Reviewed-on: https://gerrit.libreoffice.org/64837
Tested-by: Jenkins
Reviewed-by: Eike Rathke <erack@redhat.com>