tdf#99438: sc: support for .xltx export
Added support for content types for template and template
with macros (.xltm, but no user interface for this yet).

Fixed preferred extension for template: by default it is .xltx
and not .xltm (macro-enabled), because MS Excel is very strict
about matching of content-type and extension here.

Fixed wrong namespace in FilterService for ExcelFilter.

Change-Id: Ie717d31d72203601324860f069918d75082add4a
Reviewed-on: https://gerrit.libreoffice.org/65111
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>