oox smartart, accent process: fix missing bullets and large para indent
Bulleted paragraphs had a large left indent because we assumed the
bullet levels are inherited from the normal master page styles.

But that's not true, as
<https://support.office.com/en-us/article/add-bullets-to-a-shape-in-a-smartart-graphic-47edc03d-a2f8-4b54-acfd-ca103c515ab4>
points out:

"It is not possible to change the bullet style for text in a SmartArt
graphic."

This explains why the margin and bullet char info is missing from the
file format, and hints that just hardcoding these to the importer is
correct.

The result is less linebreaks in the shape text and the lost bullets are
also fixed.

Change-Id: I60bbee75f3e834551ebb1963a2f42101f3bd91d4
Reviewed-on: https://gerrit.libreoffice.org/65168
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins