sw_redlinehide_4b: visibility of flys anchored at empty nodes
The new mode would consider a fly anchored at a node that contains the
start or end of a redline but has no visible text displayed to be
hidden, while the old move-code would move it to the merged SwTextNode
and therefore keep it visible.

Change this to give the same result both on layout creation
AppendObjs() and switching AddRemoveFlysAnchoredToFrameStartingAtNode().

Reproduces with ooo108073-2.odt, tdf104735-1.odt

Change-Id: Ibac795dcf7848d78e8a2200124bc2557f35ba434
Reviewed-on: https://gerrit.libreoffice.org/65366
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>