tdf#121670 ooxmlimport: no columns in page styles, only sections
LIKELY TO EXPOSE SECTION EXPORT/IMPORT PROBLEMS. That is already
happening somewhat because support for forms/protected sections was
added in LO6.2. By making this change, it will help to expose
problems faster, with the hope that they can still be fixed for 6.2.

Columns in page styles are very problematic, because it doesn't
let you override the number of columns (except to put sub-columns
inside one of the existing columns). So, always
attempt to insert a column into it's own section,
and never into the page style itself.

I'm rather excited that this didn't cause any unit test failures.
I've made a lot of section fixes over the years
(and some this week which are required for the unit test to work).
This change seems very natural, and gets rid of a regression-prone
hack.

I found all of the existing unit tests with columns and tested them.
About 10 files - all look fine including complex files
tdf81345.docx and tdf104061_tableSectionColumns.docx

Change-Id: If02f1bfd91b1cf8210665244d0782ff926cc2869
Reviewed-on: https://gerrit.libreoffice.org/65557
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>