Data -> Statistics: sample random values in random order
The previous implementation sampled random values in the order of
the population data. This may be unexpected and is also not what
other spreadsheet implementations do. Instead, pick the random
values in random order. Keeping order can be made an option as
future feature. Code is prepared to sample WR (WithReplacement) as
well, additionally to the now (and previously) implemented WOR
(WithOutReplacement).

Change-Id: I83734d36605b28cf44c0cc2bbc2dfcafaef025f4
Reviewed-on: https://gerrit.libreoffice.org/65559
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins