tdf#47471 odfexport: use stylename if autostyle missing.
My guess is the autoStylePool can be invalidated by some events
(like clearing all attributes), and if nothing is done that
re-creates the autostyle, then Find() returns an empty string.

The normal case when all attributes are reset is that since
there are no attributes, it doesn't bother looking for the
AutoStyle, and just sticks with the given "parent" style.
However, some attributes (like lists) aren't reset, and
in those cases the autoPool didn't exist, and yet it still
looked for something, returning an empty string when
nothing was found.

If there is a style, but autostyle returns nothing, then it
should be safe to always use the given style. There are two
more areas with similar coding that I will change in separate
commits. (However, note that bug 90221 shows that an empty
autostyle might not be exactly like the given style).

Change-Id: I6a3d7b1608160f555004ca9cf1c93ec301177288
Reviewed-on: https://gerrit.libreoffice.org/65616
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>