tdf#122498 EMF+ Fix rotation of Ellipse, Arc and Pie
To draw any shape with EMF+ the shape needs to be created
and next the transformation matrix needs to be applied.
Unfortunately for Ellipse, Arc and Pie only position and ratio
was transformed. The rotation was not handled correctly.
With this patch the transformation matrix is applied after
creation of the shape.

Change-Id: I63a797e7cddfd304f94e7e5c8e2ce15063dc55e5
Reviewed-on: https://gerrit.libreoffice.org/65898
Tested-by: Jenkins
Reviewed-by: Bartosz Kosiorek <gang65@poczta.onet.pl>