writerfilter: import section margins and writingMode
Previously, only page styles got these values. But if the
section is continuous, then these properties should
be transferred to the section itself.

This patch relies on the commit for tdf#122456. It confirms
that it is a good idea for the last section to ignore
the "fake" section and keep the real section properties.
Otherwise these margins would be lost on export.

Change-Id: I45efb0d80fb9307a57ff560b2e1a26899155f827
Reviewed-on: https://gerrit.libreoffice.org/65975
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>